2. Testing/requirements.txt
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
pytest==7.4.4
pytest-cov==4.1.0
hydra-core==1.3.2
wandb==0.16.2
dvc==0.4.6
